Update opentimes.js
where you put your XML data source file (in a CORs compliant directory), e.g https://www111.lamp.ac.uk/corsdir/openingtimes.xml is key. It doesn't have to be https , that depends on you, best arrangement is to have the file on a server that serves both http and https as the protocol has to match the protocol of the web page the app is on.